+++ R/VirtualGroup.R
@@ -4,16 +4,8 @@
 #'   [NCGroup] and functions as such with the exception that it has no
 #'   associated `CFResource` and the `handle` field thus always returns `NULL`.
 #'
-#' @details This object descends from [NCGroup] and functions as such with the
-#'   exception that it has no associated `CFResource` and the `handle` field
-#'   thus always returns `NULL`.
-#'
 #' @docType class
-#' @name MemoryGroup
-NULL
-
-#' @export
-MemoryGroup <- R6::R6Class("MemoryGroup",
+VirtualGroup <- R6::R6Class("VirtualGroup",
   inherit = NCGroup,
   public = list(
     #' @description Create an instance of this class.
@@ -22,12 +14,8 @@ MemoryGroup <- R6::R6Class("MemoryGroup",
     #' @param fullname The fully qualified name of the group.
     #' @param parent The parent group of this group. The `parent` of the root
     #'   group is `NULL`.
-    #' @param title,history Title and history attributes for the group.
-    initialize = function(id, name, fullname, parent, title, history) {
+    initialize = function(id, name, fullname, parent) {
       super$initialize(id, name, fullname, parent, NULL)
-
-      self$set_attribute("title", "NC_CHAR", title)
-      self$set_attribute("history", "NC_CHAR", history)
     }
   ),
   active = list(
@@ -37,7 +25,7 @@ MemoryGroup <- R6::R6Class("MemoryGroup",
         "Group (memory)"
     },
 
-    #' @field handle Return NULL as a MemoryGroup has no resource.
+    #' @field handle Return NULL as a VirtualGroup has no resource.
     handle = function(value) {
       if (missing(value))
         NULL